requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.0
fonttools==4.54.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pybit==5.8.0
pycryptodome==3.20.0
pyparsing==3.2.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
yarl==1.12.1
